COM-3892: ulysse changes
